Extract validateCLIOptions.js
